Refactor CSV download logic: - extract a hook for the logic that converts the Plotly traces to CSV - extract a separate hook to save the CSV to file - update ChartPreview.tsx to use the updated hooks - remove unused code from ChartPreview.tsx
